+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog ##
 
+## 1.6.0
+
+- Better handle error management for mysql test connexion
+- Improve check with -1 value on PHP conf
+
+## 1.5.0
+
+- ADD: Favicon
+- ADD: WordPress Handbook link
+- UPDATE: Required version and modules
+- UPDATE: PHP requirement change from 5.6 to 7.4
+
 ## 1.4.0
 - Refactor readme
 - Branding